Add social auth buttons and componentize auth pages (#38)
