+++ web-app/.env.example
@@ -10,5 +10,19 @@ BETTER_AUTH_URL=http://localhost:3000 # Base URL of your app
 RESEND_API_KEY=<your-resend-api-key>
 
 # Social Providers
+
+# Google
 NEXT_PUBLIC_GOOGLE_CLIENT_ID=<your-google-client-id>
 GOOGLE_CLIENT_SECRET=<your-google-client-secret>
+
+# Microsoft
+NEXT_PUBLIC_MICROSOFT_CLIENT_ID=<your-microsoft-client-id>
+MICROSOFT_CLIENT_SECRET=<your-microsoft-client-secret>
+
+# Apple
+NEXT_PUBLIC_APPLE_CLIENT_ID=<your-apple-client-id>
+APPLE_CLIENT_SECRET=<your-apple-client-secret>
+
+# LinkedIn
+NEXT_PUBLIC_LINKEDIN_CLIENT_ID=<your-linkedin-client-id>
+LINKEDIN_CLIENT_SECRET=<your-linkedin-client-secret>
