--- web-app/src/app/(auth)/signin/data/index.ts
@@ -1,1 +1,0 @@
-export * from "./schema";
