--- web-app/src/app/(auth)/signup/data/index.ts
@@ -1,1 +1,0 @@
-export * from "./schema";
